+++ core/src/main/java/com/graphhopper/routing/ev/DecimalEncodedValueImpl.java
@@ -27,21 +27,13 @@ import com.graphhopper.storage.IntsRef;
  */
 public final class DecimalEncodedValueImpl extends IntEncodedValueImpl implements DecimalEncodedValue {
     private final double factor;
-    private final boolean defaultIsInfinity;
     private final boolean useMaximumAsInfinity;
 
     /**
-     * @see #DecimalEncodedValueImpl(String, int, double, double, boolean, boolean, boolean, boolean)
+     * @see #DecimalEncodedValueImpl(String, int, double, double, boolean, boolean, boolean)
      */
     public DecimalEncodedValueImpl(String name, int bits, double factor, boolean storeTwoDirections) {
-        this(name, bits, factor, false, storeTwoDirections);
-    }
-
-    /**
-     * @see #DecimalEncodedValueImpl(String, int, double, double, boolean, boolean, boolean, boolean)
-     */
-    public DecimalEncodedValueImpl(String name, int bits, double factor, boolean defaultIsInfinity, boolean storeTwoDirections) {
-        this(name, bits, 0, factor, defaultIsInfinity, false, storeTwoDirections, false);
+        this(name, bits, 0, factor, false, storeTwoDirections, false);
     }
 
     /**
@@ -50,24 +42,18 @@ public final class DecimalEncodedValueImpl extends IntEncodedValueImpl implement
      *                               maximum value.
      * @param minStorableValue       the minimum storable value. Use e.g. 0 if no negative values are needed.
      * @param factor                 the precision factor, i.e. store = (int) Math.round(value / factor)
-     * @param defaultIsInfinity      true if default should be Double.Infinity. False if 0 should be default.
      * @param negateReverseDirection true if the reverse direction should be always negative of the forward direction.
      *                               This is used to reduce space and store the value only once.
      * @param storeTwoDirections     true if forward and backward direction of the edge should get two independent values.
      * @param useMaximumAsInfinity   true if the maximum value should be treated as Double.Infinity
      */
-    public DecimalEncodedValueImpl(String name, int bits, double minStorableValue, double factor, boolean defaultIsInfinity,
+    public DecimalEncodedValueImpl(String name, int bits, double minStorableValue, double factor,
                                    boolean negateReverseDirection, boolean storeTwoDirections, boolean useMaximumAsInfinity) {
         super(name, bits, (int) Math.round(minStorableValue / factor), negateReverseDirection, storeTwoDirections);
         if (!negateReverseDirection && super.minStorableValue * factor != minStorableValue)
             throw new IllegalArgumentException("minStorableValue " + minStorableValue + " is not a multiple of the specified factor " + factor);
         this.factor = factor;
-        this.defaultIsInfinity = defaultIsInfinity;
         this.useMaximumAsInfinity = useMaximumAsInfinity;
-        if (useMaximumAsInfinity && defaultIsInfinity)
-            throw new IllegalArgumentException("defaultIsInfinity and useMaximumAsInfinity cannot be both true");
-        if (defaultIsInfinity && minStorableValue < 0)
-            throw new IllegalArgumentException("defaultIsInfinity cannot be true when minStorableValue is negative");
     }
 
     @JsonCreator(mode = JsonCreator.Mode.PROPERTIES)
@@ -85,12 +71,10 @@ public final class DecimalEncodedValueImpl extends IntEncodedValueImpl implement
                             @JsonProperty("fwd_mask") int fwdMask,
                             @JsonProperty("bwd_mask") int bwdMask,
                             @JsonProperty("factor") double factor,
-                            @JsonProperty("default_is_infinity") boolean defaultIsInfinity,
                             @JsonProperty("use_maximum_as_infinity") boolean useMaximumAsInfinity) {
         // we need this constructor for Jackson
         super(name, bits, minStorableValue, maxStorableValue, maxValue, negateReverseDirection, storeTwoDirections, fwdDataIndex, bwdDataIndex, fwdShift, bwdShift, fwdMask, bwdMask);
         this.factor = factor;
-        this.defaultIsInfinity = defaultIsInfinity;
         this.useMaximumAsInfinity = useMaximumAsInfinity;
     }
 
@@ -98,16 +82,17 @@ public final class DecimalEncodedValueImpl extends IntEncodedValueImpl implement
     public void setDecimal(boolean reverse, IntsRef ref, double value) {
         if (!isInitialized())
             throw new IllegalStateException("Call init before using EncodedValue " + getName());
-        if (Double.isInfinite(value)) {
-            if (useMaximumAsInfinity) {
+        if (useMaximumAsInfinity) {
+            if (Double.isInfinite(value)) {
                 super.setInt(reverse, ref, maxStorableValue);
                 return;
-            } else if (defaultIsInfinity) {
-                super.setInt(reverse, ref, 0);
+            } else if (value >= maxStorableValue * factor) { // equality is important as maxStorableValue is reserved for infinity
+                super.uncheckedSet(reverse, ref, maxStorableValue - 1);
                 return;
             }
+        } else if (Double.isInfinite(value))
             throw new IllegalArgumentException("Value cannot be infinite if useMaximumAsInfinity is false");
-        }
+
         if (Double.isNaN(value))
             throw new IllegalArgumentException("NaN value for " + getName() + " not allowed!");
 
@@ -123,7 +108,7 @@ public final class DecimalEncodedValueImpl extends IntEncodedValueImpl implement
     @Override
     public double getDecimal(boolean reverse, IntsRef ref) {
         int value = getInt(reverse, ref);
-        if (useMaximumAsInfinity && value == maxStorableValue || defaultIsInfinity && value == 0)
+        if (useMaximumAsInfinity && value == maxStorableValue)
             return Double.POSITIVE_INFINITY;
         return value * factor;
     }
@@ -132,7 +117,7 @@ public final class DecimalEncodedValueImpl extends IntEncodedValueImpl implement
     public double getNextStorableValue(double value) {
         if (!useMaximumAsInfinity && value > getMaxStorableDecimal())
             throw new IllegalArgumentException(getName() + ": There is no next storable value for " + value + ". max:" + getMaxStorableDecimal());
-        else if (useMaximumAsInfinity && value > getMaxStorableDecimal())
+        else if (useMaximumAsInfinity && value > (maxStorableValue - 1) * factor)
             return Double.POSITIVE_INFINITY;
         else
             return (factor * (int) Math.ceil(value / factor));
@@ -147,6 +132,7 @@ public final class DecimalEncodedValueImpl extends IntEncodedValueImpl implement
 
     @Override
     public double getMaxStorableDecimal() {
+        if (useMaximumAsInfinity) return Double.POSITIVE_INFINITY;
         return maxStorableValue * factor;
     }
 
@@ -157,8 +143,8 @@ public final class DecimalEncodedValueImpl extends IntEncodedValueImpl implement
 
     @Override
     public double getMaxOrMaxStorableDecimal() {
-        if (useMaximumAsInfinity || defaultIsInfinity)
-            throw new IllegalStateException("getMaxOrMaxStorableDecimal() is not implemented for useMaximumAsInfinity or defaultIsInfinity");
-        return getMaxOrMaxStorableInt() * factor;
+        int maxOrMaxStorable = getMaxOrMaxStorableInt();
+        if (useMaximumAsInfinity && maxOrMaxStorable == maxStorableValue) return Double.POSITIVE_INFINITY;
+        return maxOrMaxStorable * factor;
     }
 }