+++ core/src/main/java/com/graphhopper/routing/util/parsers/TurnCostParser.java
@@ -31,8 +31,6 @@ import java.util.List;
 public interface TurnCostParser {
     String getName();
 
-    void createTurnCostEncodedValues(EncodedValueLookup lookup, List<EncodedValue> registerNewEncodedValue);
-
     void handleTurnRelationTags(OSMTurnRelation turnRelation, ExternalInternalMap map, Graph graph);
 
     /**