+++ core/src/test/java/com/graphhopper/reader/osm/OSMReaderTest.java
@@ -22,6 +22,8 @@ import com.graphhopper.GHResponse;
 import com.graphhopper.GraphHopper;
 import com.graphhopper.GraphHopperTest;
 import com.graphhopper.config.Profile;
+import com.graphhopper.reader.OSMTurnRelation;
+import com.graphhopper.reader.ReaderElement;
 import com.graphhopper.reader.ReaderRelation;
 import com.graphhopper.reader.dem.ElevationProvider;
 import com.graphhopper.reader.dem.SRTMProvider;
@@ -317,26 +319,10 @@ public class OSMReaderTest {
     @Test
     public void testNegativeIds() {
         String fileNegIds = "test-osm-negative-ids.xml";
-        GraphHopper hopper = new GraphHopperFacade(fileNegIds).importOrLoad();
-        Graph graph = hopper.getBaseGraph();
-        assertEquals(4, graph.getNodes());
-        int n20 = AbstractGraphStorageTester.getIdOf(graph, 52);
-        int n10 = AbstractGraphStorageTester.getIdOf(graph, 51.2492152);
-        int n30 = AbstractGraphStorageTester.getIdOf(graph, 51.2);
-        assertEquals(GHUtility.asSet(n20), GHUtility.getNeighbors(carOutExplorer.setBaseNode(n10)));
-        assertEquals(3, GHUtility.count(carOutExplorer.setBaseNode(n20)));
-        assertEquals(GHUtility.asSet(n20), GHUtility.getNeighbors(carOutExplorer.setBaseNode(n30)));
-
-        EdgeIterator iter = carOutExplorer.setBaseNode(n20);
-        assertTrue(iter.next());
-
-        assertTrue(iter.next());
-        assertEquals(n30, iter.getAdjNode());
-        assertEquals(93147, iter.getDistance(), 1);
-
-        assertTrue(iter.next());
-        assertEquals(n10, iter.getAdjNode());
-        assertEquals(88643, iter.getDistance(), 1);
+        Exception exception = assertThrows(RuntimeException.class, () -> {
+            new GraphHopperFacade(fileNegIds).importOrLoad();
+        });
+        assertTrue(exception.getCause().getMessage().contains("Invalid OSM NODE Id: -10;"));
     }
 
     @Test
@@ -501,8 +487,8 @@ public class OSMReaderTest {
         OSMParsers osmParsers = new OSMParsers()
                 .addRelationTagParser(relConf -> new OSMBikeNetworkTagParser(bikeNetworkEnc, relConf));
         ReaderRelation osmRel = new ReaderRelation(1);
-        osmRel.add(new ReaderRelation.Member(ReaderRelation.WAY, 1, ""));
-        osmRel.add(new ReaderRelation.Member(ReaderRelation.WAY, 2, ""));
+        osmRel.add(new ReaderRelation.Member(ReaderElement.Type.WAY, 1, ""));
+        osmRel.add(new ReaderRelation.Member(ReaderElement.Type.WAY, 2, ""));
 
         osmRel.setTag("route", "bicycle");
         osmRel.setTag("network", "lcn");
@@ -590,6 +576,28 @@ public class OSMReaderTest {
 
         assertTrue(tcStorage.get(carTCEnc, edge10_11, n11, edge11_14) == 0);
         assertTrue(tcStorage.get(bikeTCEnc, edge10_11, n11, edge11_14) > 0);
+    }
+
+    @Test
+    public void testMultipleFromForNoEntry() {
+        ReaderRelation rel = new ReaderRelation(1L);
+
+        rel.setTag("restriction", "no_entry");
+        rel.add(new ReaderRelation.Member(ReaderElement.Type.WAY, 1L, "from"));
+        rel.add(new ReaderRelation.Member(ReaderElement.Type.WAY, 2L, "from"));
+        rel.add(new ReaderRelation.Member(ReaderElement.Type.NODE, 3L, "via"));
+        rel.add(new ReaderRelation.Member(ReaderElement.Type.WAY, 4L, "to"));
+
+        List<OSMTurnRelation> osmRel = OSMReader.createTurnRelations(rel);
+        assertEquals(2, osmRel.size());
+
+        assertEquals(1, osmRel.get(0).getOsmIdFrom());
+        assertEquals(4, osmRel.get(0).getOsmIdTo());
+        assertEquals(OSMTurnRelation.Type.NOT, osmRel.get(0).getRestriction());
+
+        assertEquals(2, osmRel.get(1).getOsmIdFrom());
+        assertEquals(4, osmRel.get(1).getOsmIdTo());
+        assertEquals(OSMTurnRelation.Type.NOT, osmRel.get(1).getRestriction());
     }
 
     @Test