+++ core/src/test/java/com/graphhopper/routing/util/parsers/SlopeCalculatorTest.java
@@ -75,4 +75,21 @@ class SlopeCalculatorTest {
         assertEquals(31, maxEnc.getDecimal(false, edgeRef), 1e-3);
         assertEquals(31, averageEnc.getDecimal(false, edgeRef), 1e-3);
     }
+
+    @Test
+    public void test2D() {
+        IntsRef edgeRef = new IntsRef(1);
+        PointList pointList = new PointList(5, false);
+        pointList.add(47.7283135, 11.9991135);
+        ReaderWay way = new ReaderWay(1);
+        way.setTag("point_list", pointList);
+        DecimalEncodedValue averageEnc = AverageSlope.create();
+        DecimalEncodedValue maxEnc = MaxSlope.create();
+        new EncodingManager.Builder().add(averageEnc).add(maxEnc).build();
+
+        SlopeCalculator creator = new SlopeCalculator(maxEnc, averageEnc);
+        creator.handleWayTags(edgeRef, way, IntsRef.EMPTY);
+        assertEquals(0, maxEnc.getDecimal(false, edgeRef), 1e-3);
+        assertEquals(0, averageEnc.getDecimal(false, edgeRef), 1e-3);
+    }
 }